Randomly pick first S3 host on .SO load, don't always hammer fist one
Because the .SO is reloaded and reinitted on each resource operation,
if we don't randomly pick the first host to use we'll use the 1st host
in the specified list for each and every operation.  So, we randomize
the host picked during s3Init() (mod the # of hosts, of course).