Add S3_ENABLE_MD5 option to enable upload checksums on all PUTs
New parameter, S3_ENABLE_MD5=[0|1] (default=0=off).  This enables
MD5 checksum calculations on all S3 PUT commands, setting the
Content-MD5 header.

Note that this requires reading each file effectively twice:  The
first time the file is read to calculate the initial MD5 (because
we need the MD5 to send the headers, we can't do this on the fly).
Then the second time to actually send the file over the network to
S3.

If there is a MD5 mismatch it will be logged as an error by S3 and
in rodsLog, and the iRODS system will be informed of archive failure.