Add S3_SERVER_ENCRYPT=[0|1] to enable at-rest S3 encryption
New parameter, S3_SERVER_ENCRYPT=[0|1], (default=0=off)
This sets the flag to request that S3 store the object data in encrypted
form on-disk.  This is NOT the same thing as HTTPS, and only affects the
data at-rest.  IF you need encrypted communications to S3 as well as
encryption of the object data once there, please be sure to specify
...;S3_PROTO=HTTPS;S3_SERVER_ENCRYPT=1;...
on the resource definition line.

Note also that this is not supported by some local S3 appliances or
software.  When unsupported, the S3 server will return errors for all
upload operations, logged in rodsLog.