Fix multithreaded error handling, general code cleanup
Errors in the multithreaded PUT and GET were not being properly
bumped up the chain (just reported in the logs).  The retry mechanism
was also not working because errors at any protocol callback could set
the "abort" flag, which is not right.  The callbacks should simply
set the *status to the proper error code and let the worker thread
decide if it should retry or abort the operation.

There was a memory leak in the MD5 calculation, fixed, along with
additional error checking on the Base64 conversion.

Removed the snprintf()s for error messages, following the standard
std::stringstream method that was used elsewhere.