Major rework of error/retry handling
The original code only retried certain operations, not all of
them, in the case of a retryable failure.  It also did not use the
passed-in # of retries or delay for anything other than S3 init().

This code is reworked to retry all operations that are retryable,
up to the user specified maximum, with the user specified delay
between retries.

Error handling logic in the multi-part start and end also was fixed.

It has been tested using an instrumented libs3 which returns an
retryable error randomly every ~15 operations.