Fix restarted part (start, length) for multipart PUT/GET
When a chunk of data is read or written, the multipart threads will
update the state for that specific part, incrementing the start
by the amount written/read, and decrementing the length by that same
value.

We were updating the globally set up copy of each part's (start, length).
This works as long as there are no errors in the middle of a part's
put or get.  If there is an error from S3, the whole part needs to be
re-read or re-written.  But we had already updated the part's (start, len)
tuple and would end up doing partial PUT or GETs of the part.

Now we use a thread-local copy of each part's state that is reinitted
every time a retry is attempted, so that the retried read/write will
redo the operation on the entire part.