Iterate through hostname list on retries
Instead of trying against the same S3 host IP/name when an error
occurs, iterate through the list of hosts (if present) for each
attempt.  This will allow success if some hostname/IPs in a clustered
S3 setup become unavailable.

Also explicitly set the hostname in the Stat() function.  Was not
present in original code and would default to the one that S3_Init()
was called with.