Use multipart copy operation for objects above MPU block size
Amazon objects of >5GB need to be copied using a special multipart
operation, almost identical to multipart uploads of new objects.
Override the s3PutFile to allow either putting a file or copying
an object to minimize the object impact.  Checks the object size
at the beginning of the copy and does the right thing.