DS3 WIP
Not functional and should probably be discarded. This was just a bit of code copied from
https://github.com/SpectraLogic/ds3_c_sdk/blob/master/sample/get-bucket.c
and added to the S3 resource plugin to start using the Spectra DS3 API, with the goal of gradually replacing s3 calls with ds3 calls
whenever appropriate.
The idea is to start with "list bucket" and go from there.
Spectra storage appliances supposedly honor regular S3 calls, but for actual data transfer it is recommended to use their DS3 API
to perform bulk read/write operations.