[#2053][#2056][irods/irods#6020] Fixes for test hook
Add --test, --skip-setup, and --teardown-minio options to test hook.

--skip-setup skips downloading and starting the MinIO service which
means that the user running the test assumes that MinIO is already
running in the configuration expected by the tests.

--teardown-minio will terminate the MinIO processes after running the
tests. --skip-setup and --teardown-minio are incompatible parameters
because the process object returned from subprocess.Popen must be
available to execute the termination logic, which is only the case when
setup is performed on the same test hook invocation.

Also replaced usage of deprecated features of MinIO:

- Replaced usage of MINIO_ACCESS_KEY and MINIO_SECRET_KEY with
MINIO_ROOT_USER and MINIO_ROOT_PASSWORD

- Started using --console-address to explicitly specify the API port for
the MinIO server