[irods/irods#6154] Check s3_file_write_operation error
This checks the irods::error returned by the s3_file_write_operation
function and returns early if something went wrong. If the program is
allowed to continue after a failure in this function, it can lead to a
nullptr dereference later on.

This is a minimal change required to prevent a nullptr dereference from
inside the S3 resource plugin (which causes the servicing agent to
crash). There are other places that need to prevent this kind of
situation from happening, but this is the one instance I know about.