[irods/irods#6479][irods/irods#6880] Add itouch/istream tests
Adds tests for itouch and istream into the S3 resource plugin. This
tests to make sure that the open/write/close APIs are appropriately
abstracted for dealing with different types of storage, as exposed
through these two iCommands.

These tests are minimal and likely will require expansion in the future.