[irods/irods#6154] Add tests with invalid secret key
Adds a test which makes the S3 secret key invalid and attempts an iput.
This ensures that data objects do not get stuck in a locked status even
when the S3 backend can no longer be reached by the S3 resource plugin.

The same test was created for irepl and icp as well.