[#2117][#2119] Handle timeouts on parts
Issue 2119 - Reduce the part size when flushing large files from cache
from 5 GiB to 1 GiB.

Issue 2117 - If timeouts occur on a part S3_RETRY_COUNT times, reduce
the part size in half and try the file flush again.  Continue until
either the flush succeeds or the part size is so small that you would
be over the 10,000 part limit.