[irods/irods#8554][#2271] Implemented RESOURCE_OP_READ_CHECKSUM_FROM_STORAGE_DEVICE.
The inputs (other than plugin_context) to this operation are:

- _checksum_scheme (INPUT) - The desired checksum/hash scheme.
- _returned_checksum (OUTPUT) - The returned checksum if it is available.

The implementation calls the GetObjectAttributes operation on
the object. If this returns the desired checksum/hash and the type is
"FULL_OBJECT" then _returned_checksum is populated with the value
returned from the storage device.

If the desired checksum is not avaialable an irods::error is returned.